requirements.txt
absl-py==2.1.0
bert-score==0.3.13
blosc2==2.0.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
Cython==0.29.37
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
future==1.0.0
google-auth==2.28.1
google-auth-oauthlib==1.0.0
grpcio==1.62.0
h5py==3.10.0
html5lib==1.1
huggingface-hub==0.20.3
idna==3.6
imageio==2.34.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
Jinja2==3.1.2
jiwer==3.0.3
joblib==1.3.2
kiwisolver==1.4.5
lxml==5.1.0
Markdown==3.5.2
MarkupSafe==2.1.3
matplotlib==3.7.5
mpmath==1.3.0
msgpack==1.0.7
networkx==3.0
nibabel==5.2.0
nltk==3.8.1
numexpr==2.8.6
numpy==1.24.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
oauthlib==3.2.2
packaging==23.2
pandas==2.0.3
pillow==10.2.0
protobuf==4.25.3
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycortex==1.2.7
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
shapely==2.0.3
SimpleITK==2.3.1
six==1.16.0
sympy==1.12
tables==3.8.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tokenizers==0.12.1
torch==2.2.1+cu118
torchaudio==2.2.1+cu118
torchvision==0.17.1+cu118
tornado==6.4
tqdm==4.66.2
transformers==4.22.1
triton==2.2.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
Werkzeug==3.0.1
wget==3.2
zipp==3.17.0
